Rename app in package json and Tauri conf and regenerate Android
