+++ package.json
@@ -1,5 +1,5 @@
 {
-	"name": "poc",
+	"name": "nexfeed",
 	"private": true,
 	"version": "0.0.1",
 	"type": "module",
