+++ src-tauri/gen/android/app/.gitignore
@@ -1,4 +1,4 @@
-/src/main/java/com/nexfeed/poc/generated
+/src/main/java/com/nexfeed/client/generated
 /src/main/jniLibs/**/*.so
 /src/main/assets/tauri.conf.json
 /tauri.build.gradle.kts
