+++ src-tauri/gen/android/app/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">poc</string>
-    <string name="main_activity_title">poc</string>
+    <string name="app_name">Nexfeed</string>
+    <string name="main_activity_title">Nexfeed</string>
 </resources>
