+++ src-tauri/gen/android/buildSrc/src/main/java/com/nexfeed/client/kotlin/RustPlugin.kt
